Modify. ProfileInfoData component화 한 후 중복된 syled-components 정리 후 일단은 주석처리, 나중에 서비스 시 활성화
